[AI-57] Carta compromiso

+++ app/controllers/registro_proveedores_controller.rb
@@ -171,23 +171,32 @@ class RegistroProveedoresController < ApplicationController
   #PRO-004
   def update
     @registro_proveedor = RegistroProveedor.find(params[:id])
+    puts "--------> #{@registro_proveedor.inspect}"
     asociar_institucion = @registro_proveedor.asociar_institucion
     contribuyente_id = @registro_proveedor.contribuyente_id
     rut_institucion = @registro_proveedor.rut_institucion
     registro_proveedores_params_sin_editar = registro_proveedores_params.except(:editar)
-
+    puts "---------------->2 #{registro_proveedores_params_sin_editar[:carta_compromiso].nil?}"
+    puts "aaa---> #{request.format.symbol}"
     respond_to do |format|
       if @registro_proveedor.update(registro_proveedores_params_sin_editar)
-        RegistroProveedor::UpdateService.new(@registro_proveedor, registro_proveedores_params, asociar_institucion, contribuyente_id, rut_institucion).perform
-        if params[:registro_proveedor][:editar] != 'true'
-          flujo = Flujo.where(id: 1000, contribuyente_id: 1000, tipo_instrumento_id: 26).first_or_create
-          tarea_pendiente = TareaPendiente.where(flujo_id: flujo.id, user_id: current_user.id).first
-          tarea_pendiente.destroy
-          tarea = Tarea.where(nombre: "PRO-003").first
-          TareaPendiente.create(flujo_id: flujo.id, tarea_id: tarea.id, estado_tarea_pendiente_id: EstadoTareaPendiente::NO_INICIADA, user_id: @registro_proveedor.user_encargado, data: @registro_proveedor.id)
-        end
+        if registro_proveedores_params_sin_editar[:carta_compromiso].nil?
+          RegistroProveedor::UpdateService.new(@registro_proveedor, registro_proveedores_params, asociar_institucion, contribuyente_id, rut_institucion).perform
+       
+          if params[:registro_proveedor][:editar] != 'true'
+            flujo = Flujo.where(id: 1000, contribuyente_id: 1000, tipo_instrumento_id: 26).first_or_create
+            tarea_pendiente = TareaPendiente.where(flujo_id: flujo.id, user_id: current_user.id).first
+            tarea_pendiente.destroy
+            tarea = Tarea.where(nombre: "PRO-003").first
+            TareaPendiente.create(flujo_id: flujo.id, tarea_id: tarea.id, estado_tarea_pendiente_id: EstadoTareaPendiente::NO_INICIADA, user_id: @registro_proveedor.user_encargado, data: @registro_proveedor.id)
+          end
+       end
         format.js {
           render js: "window.location='#{root_path}'"
+          flash[:success] = "Registro enviado correctamente"
+        }
+        format.html {
+          redirect_to root_path
           flash[:success] = "Registro enviado correctamente"
         }
        
@@ -431,6 +440,13 @@ class RegistroProveedoresController < ApplicationController
   def evaluacion_proveedores
   end
 
+  #PRO-010
+  def enviar_carta_compromiso
+    @registro_proveedor = RegistroProveedor.find(params[:id])
+    @descargables_tarea = DescargableTarea.where(tarea_id: 104)
+
+  end
+
   def descargar_documentos_proveedores
     require 'zip'
     archivo_zip = Zip::OutputStream.write_buffer do |stream|
@@ -517,7 +533,7 @@ class RegistroProveedoresController < ApplicationController
   def registro_proveedores_params
     params.require(:registro_proveedor).permit(:rut, :nombre, :apellido, :email, :telefono, :profesion, :direccion, :region, :comuna, :ciudad, :asociar_institucion, :tipo_contribuyente_id, :terminos_y_servicion,
       :rut_institucion, :nombre_institucion, :tipo_contribuyente, :editar, :tipo_proveedor_id, :direccion_casa_matriz, :region_casa_matriz, :comuna_casa_matriz, :ciudad_casa_matriz, :contribuyente_id, :respuesta_comentario,
-      :archivo_respuesta_rechazo, :comentario_directiva, :respuesta_comentario_directiva, :archivo_respuesta_rechazo_directiva, :fecha_aprobado, :fecha_actualizado, :archivo_aprobado_directiva, :archivo_aprobado_directiva_cache,
+      :archivo_respuesta_rechazo, :comentario_directiva, :respuesta_comentario_directiva, :archivo_respuesta_rechazo_directiva, :fecha_aprobado, :fecha_actualizado, :archivo_aprobado_directiva, :archivo_aprobado_directiva_cache, :carta_compromiso, :carta_compromiso_cache,
       certificado_proveedores_attributes: [:id, :materia_sustancia_id, :actividad_economica_id,  :archivo_certificado, :archivo_certificado_cache, :_destroy], documento_registro_proveedores_attributes: [:id, :description, :archivo, :archivo_cache, :_destroy],
       certificado_proveedor_extras_attributes: [:id, :materia_sustancia_id, :actividad_economica_id, :archivo, :archivo_cache, :_destroy], documento_proveedor_extras_attributes: [:id, :description, :archivo, :archivo_cache, :_destroy])
   end
